+++ include/smokerand/hwtests.h
@@ -62,7 +62,7 @@ typedef struct {
 
 typedef struct {
     unsigned long long nvalues; ///< Number of pseudorandom values processed by the test.
-    int nlevels; ///< Number of blocks levels.
+    unsigned int nlevels; ///< Number of blocks levels.
 } HammingDistrOptions;